requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.17,<4.0
django-cors-headers>=4.9,<5.0
openai>=2.32,<3.0
python-dotenv>=1.0,<2.0
